==> R markdown analysis/01_nuc_20250922.Rmd <==
---
title: "HBV"
output: html_document
date: "2025-09-22"
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(tidyverse)
library(readxl)
library(openxlsx)

library(stringi)

library(ggtext)
library(patchwork)

library(scales)
```

```{r}
# create folder for each file under year

dir.create("../df", showWarnings = TRUE, recursive = FALSE, mode = "0777")

dir.create("../df/analysis", showWarnings = TRUE, recursive = FALSE, mode = "0777")
```

```{r}
# create folder for each file under year

dir.create("../plot", showWarnings = TRUE, recursive = FALSE, mode = "0777")
```

```{r}
# create folder for each file under year

dir.create("../table", showWarnings = TRUE, recursive = FALSE, mode = "0777")
```


```{r}
list.files("../rds")
```

```{r}
# read RDS

rds_name <- list.files("../rds") %>% str_subset("RDS") %>% str_remove("\\.RDS")

rds_name
```

```{r}
rds_path <- paste0("../rds/", rds_name, ".RDS")
```

```{r}
for (i in 1:length(rds_name)){
  assign(rds_name[i]　, readRDS(rds_path[i]))
}
```



```{r}
df_oral_nuc <- df_oral %>%
  filter(str_detect(DRUG_name, "バラクルード|エンテカビル|テノゼット|ベムリディ"))
```


```{r}
df_oral_nuc <- df_oral_nuc %>%
  mutate(NUC = case_when(str_detect(DRUG_name, "バラクルード")　 ~ "ETV [branded]",
                         str_detect(DRUG_name, "エンテカビル")　 ~ "ETV [generic]",
                         str_detect(DRUG_name, "テノゼット")　 ~ "TDF",
                         str_detect(DRUG_name, "ベムリディ")　 ~ "TAF"))　%>%
  distinct(NUC, COST_per) %>%
  mutate(COST_per = as.numeric(COST_per))

df_oral_nuc
```


```{r}
df_oral_nuc <- df_oral_nuc %>%
  mutate(NUC = case_when(NUC == "ETV [generic]" & COST_per == 220.8	 ~ "ETV [generic 1]",
                         NUC == "ETV [generic]" & COST_per == 89.0	 ~ "ETV [generic 2]",
                         .default = NUC))

df_oral_nuc
```

```{r}
df_oral_nuc <- df_oral_nuc %>%
  mutate(YEAR = "2023") %>%
  mutate(USD_JPY = 140.62)

df_oral_nuc
```

```{r}
print(10000/140.62)

print(20000/140.62)
```



```{r}
df_oral_nuc <- df_oral_nuc %>%
  mutate(COST_jp = COST_per *365) %>%
  mutate(COST_usd = COST_jp/USD_JPY) %>%
  mutate(IRP = case_when(str_detect(NUC, "ETV") ~ 36,
                         str_detect(NUC, "TDF") ~ 30)) %>%
  mutate(MPR = COST_usd/IRP) 

df_oral_nuc
```

```{r}
df_oral_nuc <- df_oral_nuc %>%
  mutate(FILL = case_when(str_detect(NUC, "ETV") ~ "ETV",
                         str_detect(NUC, "TDF") ~ "TDF",
                         .default = "TAF")) %>%
  mutate(FILL = factor(FILL, levels = c("ETV", "TDF", "TAF"))) %>%
  mutate(ALPHA = case_when(str_detect(NUC, "TDF|1\\]|TAF") ~ "a",
                           .default = "b"))
```

```{r}
df_oral_nuc <- df_oral_nuc %>%
  mutate(NUC = str_replace(NUC, "V\\s\\[", "V\n\\["))

df_oral_nuc 
```


```{r}
levels_nuc <- df_oral_nuc %>%
  arrange(COST_usd) %>%
  pull(NUC)

levels_nuc
```

```{r}
df_oral_nuc <- df_oral_nuc %>% mutate(NUC = factor(NUC, levels = levels_nuc))
```


```{r}
df_oral_nuc
```

```{r}
1004/140.62
```
```{r}
1004 * 40* 52

1004/140.62 * 40* 52
```


```{r}
df_oral_nuc <- df_oral_nuc %>%
  mutate(AFF = COST_jp*100/(1004* 40 * 52) ) %>%
  mutate(AFF_1 = AFF * 0.1) %>%
  mutate(AFF_2 = AFF * 0.2) %>%
  mutate(AFF_3 = AFF * 0.3)
```


```{r}
#p1_a1 <- 
  df_oral_nuc %>%
  mutate(COST_per = COST_per/USD_JPY) %>%
  ggplot(aes(COST_per, fct_reorder(NUC, COST_per), fill = FILL, alpha = ALPHA))+
  geom_col(width = 0.7)+
  scale_fill_manual(values = c("#2ca02c",
                               "#ff7f0e",
                               "#9467bd"))+
  scale_alpha_manual(values = c(1, 0.5))+
  scale_x_continuous(expand = c(0, 0),
                     limits = c(0, 8),
                     breaks = seq(0, 8, 2))+
  geom_text(aes(x = COST_per + 8/25, label = round(COST_per, digits = 1)),
            hjust = 0, size = 2.8, alpha = 1, colour = "black")+
  coord_cartesian(clip = "off")+
  labs(tag = "a", 
       x = "Daily cost (USD)",
       y = "Drug")+
  theme_classic()+
  theme(plot.tag = element_text(size = 12, colour = "black", face = "bold"),
        axis.title = element_text(size = 8, colour = "black", face = "bold"),
        axis.line = element_line(linewidth = 0.3, colour = "black"),
        axis.ticks = element_line(linewidth = 0.3, colour = "black"),
        axis.text = element_text(size = 8, colour = "black"),
        legend.position = "none",
        plot.margin = margin(0, 5 , 0, 0, "mm"))

#p1_a1
```


```{r}
df_oral_nuc 
```


```{r}
p1_a1 <- df_oral_nuc %>%
  ggplot(aes(COST_usd, fct_reorder(NUC, COST_usd), fill = FILL, alpha = ALPHA))+
  geom_col(width = 0.5)+
  scale_fill_manual(values = c("#2ca02c",
                               "#ff7f0e",
                               "#9467bd"))+
  scale_alpha_manual(values = c(1, 0.5))+
  scale_x_continuous(expand = c(0, 0),
                     limits = c(0, 3000),
                     breaks = seq(0, 3000, 1000),
                     labels = label_comma())+
  geom_text(aes(x = COST_usd + 3000/25, label = round(COST_usd) %>% scales::comma()),
            hjust = 0, size = 2.8, alpha = 1, colour = "black")+
  coord_cartesian(clip = "off")+
  labs(tag = "a", 
       x = "Annual cost (USD)",
       y = "Drug")+
  theme_classic()+
  theme(plot.tag = element_text(size = 12, colour = "black", face = "bold"),
        axis.title = element_text(size = 8, colour = "black", face = "bold"),
        axis.text.x = element_text(size = 8, colour = "black"),
        axis.text.y = element_text(size = 8, colour = "black"),
                                      # lineheight = 1.2),
        axis.line = element_line(linewidth = 0.3, colour = "black"),
        axis.ticks = element_line(linewidth = 0.3, colour = "black"),
        legend.position = "none",
        plot.margin = margin(0, 7.5 , 0, 0, "mm"))

p1_a1
```

```{r}
df_oral_nuc
```

```{r}
label_mpr <- df_oral_nuc %>%
   mutate(x =  case_when(NUC =="TAF" ~ 30,
                           .default = MPR + 60/25)) %>%
  mutate(label = round(MPR, digits = 1)) %>%
  mutate(label = case_when(NUC =="TAF" ~"Not available",
                           .default = as.character(label))) %>%
  mutate(hjust = case_when(NUC =="TAF" ~ 0.5,
                           .default = 0))
```



```{r}
p1_a2 <- df_oral_nuc %>%
  mutate(MPR = replace_na(MPR, 0)) %>%
  ggplot(aes(MPR, NUC, fill = FILL, alpha = ALPHA))+
  geom_col(width = 0.5)+
  scale_fill_manual(values = c("#2ca02c",
                               "#ff7f0e",
                               "#9467bd"))+
  scale_alpha_manual(values = c(1, 0.5))+
  scale_x_continuous(expand = c(0, 0),
                     limits = c(0, 60),
                     breaks = seq(0, 60, 20))+
    geom_text(data = label_mpr,
              aes(x = x , y = NUC, label = label, hjust = hjust),
              size = 2.8, alpha = 1, colour = "black" ,
              inherit.aes = FALSE)+
   coord_cartesian(clip = "off")+
  labs(x = "Price ratio",
       y = "Drug")+
  theme_classic()+
  theme(axis.title.x = element_text(size = 8, colour = "black", face = "bold"),
        axis.title.y = element_blank(),
        axis.text.x = element_text(size = 8, colour = "black"),
        axis.text.y = element_blank(),
        axis.line = element_line(linewidth = 0.3, colour = "black"),
        axis.ticks = element_line(linewidth = 0.3, colour = "black"),
        legend.position = "none",
        plot.margin = margin(0, 7.5 , 0, 0, "mm"))

p1_a2
```


```{r}
df_oral_nuc_long <- df_oral_nuc %>%
   pivot_longer(cols = starts_with("AFF"),
               names_to = "AFF_cat",
               values_to = "AFF_value")


df_oral_nuc_long
```


```{r}
label_nuc_aff <- df_oral_nuc %>%
  mutate(label_1 = round(AFF, digits = 2) %>% format(nsmall = 2) %>% str_remove_all("\\s")) %>%
  mutate(label_2 = paste0(round(AFF_1, digits = 2) %>% format(nsmall = 2) %>%
                            str_remove_all("\\s"),
                        " - ",
                        round(AFF_3, digits = 2) %>% format(nsmall= 2) %>% str_remove_all("\\s"))) %>%
  arrange(desc(NUC)) %>%
  bind_cols(tibble(y1 = c(5:1) - 0.3)) %>%
  mutate(x1 = AFF) %>%
  mutate(hjust1 = 0.5) %>%
  bind_cols(tibble(y2 = c(5:1) + 0.3)) %>%
  mutate(x2 = case_when(str_detect(NUC, "ETV") & str_detect(NUC, "generic") ~ 0,
                       .default = AFF_2)) %>%
  mutate(hjust2 = case_when(str_detect(NUC, "ETV") & str_detect(NUC, "generic") ~ 0,
                       .default = 0.5))

label_nuc_aff
```


```{r}
# affordability

p1_a3 <- df_oral_nuc %>%
   ggplot()+
#    geom_segment(aes(x = AFF_1, xend = AFF_3,
#                   y = NUC, yend = NUC), linewidth = 0.4) +
  geom_point(data = df_oral_nuc_long %>% filter(AFF_cat %in% c("AFF")),
             aes(x = AFF_value, y = NUC, colour = AFF_cat),
             size = 1,
             inherit.aes = FALSE)+
  scale_colour_manual(values = c("#1f77b4", "#d62728"),
                    name = "Cost",
                    labels = c("Original",
                               "Out-of-pocket"))+
    geom_text(data = label_nuc_aff,
            aes(x = x1, y = y1, label = label_1, hjust = hjust1),
             vjust = 0.5, size = 2.8, colour = "black",
            inherit.aes = FALSE)+
#  geom_text(data = label_nuc_aff,
#            aes(x = x2, y = y2, label = label_2, hjust = hjust2),
#             vjust = 0.5, size = 2.8, colour = "black",
#            inherit.aes = FALSE)+
  scale_x_continuous(breaks = seq(0, 20, 5),
                     limits = c(0, 20))+
  coord_cartesian(clip = "off")+
  labs(
       x = "Affordability (%)",
       y = "Drug")+
  guides(colour = guide_legend(override.aes = list(size = 2.6,
                                                   fill=NA)))+
  theme_classic()+
  theme(plot.title = element_text(size = 12, colour = "black", face = "bold"),
        axis.title.x = element_text(size = 8, colour = "black", face = "bold"),
        axis.title.y = element_blank(),
        axis.text.x = element_text(size = 8, colour = "black"),
        axis.text.y = element_blank(),
        axis.line = element_line(linewidth = 0.3, colour = "black"),
        axis.ticks = element_line(linewidth = 0.3, colour = "black"),
        legend.title = element_blank(),
          #element_text(size = 8, colour = "black", face = "bold"),
        legend.text = element_text(size = 8, colour = "black",
                                   margin = margin(0, 0, 0, 0, unit = "mm")),
        legend.position = "inside",
        legend.position.inside = c(0.55, 0),
        legend.justification.inside = c(0, 0),
        legend.background = element_blank(),
        legend.box.background = element_blank(),
    #    legend.margin = margin(0, 0, 0, 0),
      legend.spacing.x = unit(0, "mm"),
      legend.spacing.y = unit(0, "mm"),
      legend.key = element_rect(fill = NA),
        plot.margin = margin(0, 0 , 0, 0, "mm"))

 p1_a3
```

```{r}
p1_a <- p1_a1 + p1_a2 + p1_a3 +
  plot_layout(ncol = 3, widths = c(2, 2, 5), axis_titles = "collect_y")

p1_a 
```

```{r}
saveRDS(p1_a1, "../plot/p1_a1.RDS")
saveRDS(p1_a2, "../plot/p1_a2.RDS")
saveRDS(p1_a3, "../plot/p1_a3.RDS")
```



# create table


```{r}
df_oral_table <- df_oral %>%
  filter(str_detect(DRUG_name, "バラクルード|エンテカビル|テノゼット|ベムリディ")) %>%
  mutate(NUC = case_when(str_detect(DRUG_name, "バラクルード")　 ~ "ETV (brand)",
                         str_detect(DRUG_name, "エンテカビル")　 ~ "ETV (generic)",
                         str_detect(DRUG_name, "テノゼット")　 ~ "TDF",
                         str_detect(DRUG_name, "ベムリディ")　 ~ "TAF"))　%>%
  distinct(DRUG_name, DRUG_code, COST_code, .keep_all = TRUE)

df_oral_table
```
