==> R markdown analysis/02_monioring_20250922.Rmd <==
---
title: "HBV"
output: html_document
date: "2025-09-22"
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(RCurl)
library(tidyverse)
library(XML)
library(xml2)
library(rvest)
library(readxl)
library(openxlsx)

library(stringi)

library(patchwork)
```


```{r}
list.files("../rds")
```

```{r}
# read RDS

rds_name <- list.files("../rds") %>% str_subset("RDS") %>% str_remove("\\.RDS")

rds_name
```

```{r}
rds_path <- paste0("../rds/", rds_name, ".RDS")
```

```{r}
for (i in 1:length(rds_name)){
  assign(rds_name[i]　, readRDS(rds_path[i]))
}
```


```{r}
df_combine_add %>%
  mutate(across(c(POINT, TOTAL), ~ na_if(.x, "-")))
```


```{r}
rds_list <- map(rds_path, function(x){
  
  if( sum(colnames(readRDS(x)) %in% "POINT") == 1){
    readRDS(x) %>%
      mutate(across(c(POINT, TOTAL), ~ na_if(., "-"))) %>%
      mutate(across(c(POINT, TOTAL), ~ replace_na(., "0"))) %>%
    mutate(across(c(POINT, TOTAL), ~ as.numeric(.x))) %>%
    return()
  }  
  
  else{
    readRDS(x) %>%
      mutate(across(c(COST_per, TOTAL), ~ na_if(.x, "-"))) %>%
      mutate(across(c(COST_per, TOTAL), ~ replace_na(., "0"))) %>%
    mutate(across(c(COST_per, TOTAL), ~ as.numeric(.x))) %>%
    return()
  }
})
```


```{r}
for (i in 1:length(rds_name)){
  assign(rds_name[i]　, rds_list[[i]])
}
```


# read df_oral_nuc

```{r}
df_oral_nuc <- readRDS("../df/analysis/df_oral_nuc.RDS")
```


```{r}
df_combine_reg %>% 
  filter(str_detect(PRO_name, "外来診療料|処方") &
           SHEET == "out") %>% View()
```


```{r}
# visit
# hospital with > 200 beds

df_clinic <- df_combine_reg %>% 
  filter(PRO_name %in% c("外来診療料", "処方箋料（リフィル以外・その他）") &
           SHEET == "out") %>%
  mutate(PRO = "Clinic visit")

df_clinic 
```

```{r}
df_combine_reg %>% 
  filter(PRO_name == "外来診療料")
```



```{r}
# prescription + pharmacy
df_combine_reg %>%
  filter(str_detect(PRO_name, "処方料|調剤料"))　%>% View()
```

```{r}
# assume outpatient prescription at other pharmacy

df_combine_reg %>%
  filter(FOLDER == "pharmacy") %>% View()


```

```{r}
df_combine_reg %>% count(FOLDER)
```


```{r}
df_combine_reg %>%
  filter(FOLDER == "pharmacy") %>% 
  group_by(CLASS_name) %>% 
  slice_max(TOTAL, n =1) %>%
  View()
```

```{r}
df_pharma <- df_combine_reg %>%
  filter(FOLDER == "pharmacy") %>% 
  group_by(CLASS_name) %>% 
  slice_max(TOTAL, n =1) %>%
  mutate(PRO = "Pharmacy")
```




```{r}
df_pharma
```


```{r}
df_combine_reg %>% 
#  mutate(PRO_name_half = stri_trans_general(PRO_name, "fullwidth-halfwidth")) %>%
  filter(str_detect(PRO_name, "血液")) %>% View()


# test
# 末梢血液一般検査

# procedure
# 血液採取（静脈）
# 血液学的検査判断料
# 生化学的検査（１）判断料
```

```{r}
df_combine_reg %>% 
#  mutate(PRO_name_half = stri_trans_general(PRO_name, "fullwidth-halfwidth")) %>%
  filter(str_detect(PRO_name, "悪性腫瘍特異物質治療管理料")) %>% View()
```


```{r}
#blood tests

# CBC (for platelet)
# AST, ALT
# Cr (for dose adjustment)
# PT, Alb, T-Bil
# HBsAg, HBV-DNA

df_combine_reg %>% 
  mutate(PRO_name_half = stri_trans_general(PRO_name, "fullwidth-halfwidth")) %>%
  filter(str_detect(PRO_name_half, "HBs|HBV|AST|ALT|PIVKA")) %>% View()

# blood test
# 末梢血液一般検査
# アスパラギン酸アミノトランスフェラーゼ（ＡＳＴ）
# アラニンアミノトランスフェラーゼ（ＡＬＴ）
# クレアチニン
# プロトロンビン時間（ＰＴ）
# アルブミン（ＢＣＰ改良法・ＢＣＧ法）
# 総ビリルビン
# ＨＢｓ抗原
# ＨＢＶ核酸定量
# α－フェトプロテイン（ＡＦＰ）
# ＰＩＶＫＡ－２定量

# procedure
# 血液採取（静脈）
# 血液学的検査判断料
# 生化学的検査（１）判断料
# 悪性腫瘍特異物質治療管理料（その他のもの）（２項目以上） exclude
```

```{r}
blood_reg <- c("末梢血液一般検査",
                "アスパラギン酸アミノトランスフェラーゼ（ＡＳＴ）",
                "アラニンアミノトランスフェラーゼ（ＡＬＴ）",
                "クレアチニン",
                "プロトロンビン時間（ＰＴ）",
                "アルブミン（ＢＣＰ改良法・ＢＣＧ法）",
                "総ビリルビン",
                "ＨＢｓ抗原",
                "ＨＢＶ核酸定量",
                "α－フェトプロテイン（ＡＦＰ）",
                "ＰＩＶＫＡ－２定量",
                "血液採取（静脈）",
                "血液学的検査判断料",
                "生化学的検査（１）判断料"
                )
```

```{r}
df_blood <- df_combine_reg %>%
  filter(PRO_name %in% blood_reg & SHEET == "out") %>%
  mutate(PRO = "Blood test") 
```


```{r}
df_combine_reg %>% 
  filter(str_detect(PRO_name, "アルブミン")) %>% View()
```

```{r}
# imaging test reg

# CT and MRI choose the most common one

# US
# 超音波検査（断層撮影法（心臓超音波検査を除く。））（その他の場合）（胸腹部）

# CT
# ＣＴ撮影（１６列以上６４列未満のマルチスライス型の機器による場合）（一連につき）

# MRI
# 磁気共鳴コンピューター断層撮影（ＭＲＩ撮影）（１．５テスラ以上３テスラ未満の機器による場合）（一連につき）

df_combine_reg %>%
  filter(str_detect(PRO_name, "超音波|ＣＴ|ＭＲＩ")) %>% View()
```

```{r}
# radiology add
# 造影剤使用加算（ＣＴ撮影）
# 造影剤使用加算（ＭＲＩ撮影）
# 電子画像管理加算（コンピューター断層診断料）（一連につき）
# 画像診断管理加算２（コンピューター断層診断）

df_combine_add %>%
  filter(str_detect(PRO_name, "コンピューター断層診断料")) %>% View()
```


```{r}
df_combine_add %>%
  filter(str_detect(PRO_name, "画像診断管理加算")) %>% View()

# 画像診断管理加算
```

```{r}
# US

df_us <- df_combine_reg %>%
  filter(str_detect(PRO_name, "超音波検査（断層撮影法（心臓超音波検査を除く。））（その他の場合）（胸腹部）")　&
           SHEET == "out") %>%
  mutate(PRO = "Ultrasound")
  
df_us 
```

```{r}
#

df_combine_reg %>%
  filter(str_detect(PRO_name, "エラストグラフィ|肝硬度"))
```

```{r}
df_elasto <- df_combine_reg %>%
  filter(str_detect(PRO_name, "エラストグラフィ|肝硬度")) %>%
  distinct(POINT, .keep_all = TRUE) %>%
  mutate(PRO = "Transient\nelastography") 

df_elasto
```


```{r}
df_combine_reg %>%
  filter(str_detect(PRO_name, "ＣＴ撮影") & 
           !str_detect(PRO_name, "頭")　&
            !str_detect(PRO_name, "脳")　&
           SHEET == "out") %>% View() 
```

```{r}
df_ct_reg <- df_combine_reg %>%
  filter(PRO_name == "ＣＴ撮影（１６列以上６４列未満のマルチスライス型の機器による場合）（一連につき）"　& SHEET == "out") %>% 
  mutate(PRO = "CT") 

df_ct_reg 
```
```{r}
df_combine_add %>%
  filter(str_detect(PRO_name, "造影剤使用加算（ＭＲＩ撮影）|電子画像管理加算|画像診断管理加算")) %>%
  filter(str_detect(SHEET, "out")) %>%
  View()

# 電子画像管理加算（コンピューター断層診断料）（一連につき）
# 画像診断管理加算２（コンピューター断層診断）
# 造影剤使用加算（ＭＲＩ撮影）

```

```{r}
df_ct_add <- df_combine_add %>%
  filter(PRO_name %in% c("電子画像管理加算（コンピューター断層診断料）（一連につき）",
                         "画像診断管理加算２（コンピューター断層診断）",
                         "造影剤使用加算（ＭＲＩ撮影）") &
           SHEET =="out_add") %>%
  mutate(PRO = "CT") 

df_ct_add
```
```{r}
# get contrast medium

#df_ct_cm <- 
  df_inject %>% filter(CLASS_name == "Ｘ線造影剤") %>% group_by(DRUG_name) %>% View()
#  summarise(sum = sum(TOTAL)) %>%
#  arrange(desc(sum)) 
```



```{r}
df_ct_cm <- df_inject %>% 
  filter(DRUG_name == "イオパミロン注３７０シリンジ　７５．５２％１００ｍＬ" & SHEET == "outpt_inpr") %>%
  mutate(PRO = "CT") %>%
  mutate(POINT = COST_per/10)
```

```{r}
df_combine_reg %>%
  filter(str_detect(PRO_name, "ＭＲＩ")) %>%
  filter(str_detect(SHEET, "out")) %>%
  View()
```


```{r}
# CEMRI

df_combine_add %>%
  filter(str_detect(PRO_name, "ＭＲＩ")) %>%
  filter(str_detect(SHEET, "out")) %>%
  View()

#磁気共鳴コンピューター断層撮影（ＭＲＩ撮影）（１．５テスラ以上３テスラ未満の機器による場合）（一連につき）

# 造影剤使用加算（ＭＲＩ撮影）
# 電子画像管理加算（コンピューター断層診断料）（一連につき）
# 画像診断管理加算２（コンピューター断層診断）

```


```{r}
df_mri_reg <- df_combine_reg %>%
  filter(str_detect(PRO_name,
                    "磁気共鳴コンピューター断層撮影（ＭＲＩ撮影）（１．５テスラ以上３テスラ未満の機器による場合）（一連につき）") & 
           SHEET == "out") %>% 
  mutate(PRO = "MRI")

df_mri_reg
```

```{r}
df_mri_add <- df_combine_add %>%
  filter(PRO_name %in% c("造影剤使用加算（ＭＲＩ撮影）",
                         "電子画像管理加算（コンピューター断層診断料）（一連につき）",
                         "画像診断管理加算２（コンピューター断層診断）") &
           SHEET == "out_add") %>% 
  mutate(PRO = "MRI")

df_mri_add
```

```{r}
  df_inject %>% filter(str_detect(CLASS_name, "その他の診断用薬"))
```


```{r}
  df_inject %>% filter(str_detect(DRUG_name, "プリモビスト")) %>% View()

```

```{r}
df_mri_cm <-   df_inject  %>% filter(str_detect(DRUG_name, "ＥＯＢ・プリモビスト注シリンジ　１８．１４３％１０ｍＬ") &
                      str_detect(SHEET, "out")) %>% 
  mutate(PRO = "MRI") %>%
  mutate(POINT = COST_per/10)
  
  #ＥＯＢ・プリモビスト注シリンジ　１８．１４３％１０ｍＬ

```


```{r}
df_test <- df_clinic %>%
  bind_rows(df_pharma) %>%
  bind_rows(df_blood) %>%
  bind_rows(df_us) %>%
  bind_rows(df_elasto) %>%
  bind_rows(df_ct_reg) %>%
  bind_rows(df_ct_add) %>%
  bind_rows(df_ct_cm) %>%
  bind_rows(df_mri_reg) %>%
  bind_rows(df_mri_add) %>%
  bind_rows(df_mri_cm )
```

```{r}
#test_levels <- df_test %>% pull(PRO) %>% unique() %>% rev()

#test_levels
```

```{r}
df_test <- df_test %>%
  mutate(PRO_name = case_when(is.na(PRO_name) ~ DRUG_name,
                              .default = PRO_name))
```

```{r}
df_test %>%
  group_by(PRO_name, PRO) %>%
  filter(n()>1)
```


```{r}
df_test %>%
  group_by(PRO_name, PRO) %>%
  filter(n()>1)

# zero
```


```{r}
df_test <- df_test %>%
  mutate(USD_JPY = 140.62) %>%
  mutate(COST_jp = POINT * 10)
```

```{r}
df_test_cost <- df_test %>%
  group_by(PRO, USD_JPY) %>%
  summarise(COST_jp = sum(COST_jp))

df_test_cost
```

```{r}
df_test_cost <- df_test_cost %>%
  mutate(COST_usd = COST_jp/USD_JPY) %>%
  mutate( AFF = COST_jp*100/(1004*40*52) ) %>%
  mutate(AFF_1  = AFF * 0.1)%>%
  mutate(AFF_2 = AFF*0.2) %>%
  mutate(AFF_3 = AFF*0.3)
```

```{r}
levels_pro <- df_test_cost %>% arrange(COST_jp) %>% pull(PRO)

levels_pro                                                    
```
```{r}
df_test_cost <- df_test_cost %>%
  mutate(PRO = factor(PRO, levels = levels_pro))
```

```{r}
df_test_cost
```


```{r}
p1_b1 <- df_test_cost %>%
  ggplot(aes(COST_usd, PRO))+
  geom_col(width = 0.5,
           fill = "#7f7f7f")+
  geom_text(aes(x = COST_usd + 300/50, y = PRO,
                label = round(COST_usd, digit = 2) %>%
                  format(nsmall = 2) %>%
                  str_remove_all("\\s")),
            hjust = 0, size = 2.8, colour = "black")+
  labs(tag = "b",
       x = "Original cost (USD)",
       y = "Procedure")+
  scale_x_continuous(expand = c(0, 0),
                     limits = c(0, 300),
                     breaks = seq(0, 300, 100))+
  coord_cartesian(clip = "off") +
  theme_classic()+
   theme(plot.tag = element_text(size = 12, colour = "black", face = "bold"),
         axis.title = element_text(size = 8, colour = "black", face = "bold"),
        axis.text = element_text(size = 8, colour = "black"),
        axis.line = element_line(linewidth = 0.3, colour = "black"),
        axis.ticks = element_line(linewidth = 0.3, colour = "black"),
        legend.position = "none",
        plot.margin = margin(0, 7.5 , 0, 0, "mm"))

p1_b1
```

```{r}
df_test_cost %>%
  ggplot(aes(COST_usd, PRO))+
  geom_col(width = 0.5,
           fill = "#7f7f7f")+
  geom_text(aes(x = COST_usd, y = PRO,
                label = round(COST_usd, digit = 2)),
            hjust = 0, size = 2.8, colour = "black")+
  labs(tag = "b",
       x = "Original cost (USD)",
       y = "Procedure")+
  scale_x_continuous(expand = c(0, 0),
                     limits = c(0, 300),
                     breaks = seq(0, 300, 100))+
  coord_cartesian(clip = "off") +
  theme_classic()+
   theme(plot.tag = element_text(size = 12, colour = "black", face = "bold"),
         axis.title = element_text(size = 8, colour = "black", face = "bold"),
        axis.text = element_text(size = 8, colour = "black"),
        axis.line = element_line(linewidth = 0.3, colour = "black"),
        axis.ticks = element_line(linewidth = 0.3, colour = "black"),
        legend.position = "none",
        plot.margin = margin(0, 7.5 , 0, 0, "mm"))
```



```{r}
df_test_cost_long <- df_test_cost %>%
   pivot_longer(cols = starts_with("AFF"),
               names_to = "AFF_cat",
               values_to = "AFF_value")


df_test_cost_long 
```


```{r}
label_test_cost <- df_test_cost %>%
  mutate(label_1 = round(AFF, digits = 2) ) %>%
  mutate(label_2 = paste0(round(AFF_1, digits = 2),
                        " - ",
                        round(AFF_3, digits = 2) )) %>%
  arrange(desc(PRO)) %>%
  bind_cols(tibble(y1 = c(7:1) - 0.3)) %>%
  mutate(x1 = AFF) %>%
  mutate(hjust1 = case_when(PRO %in% c("MRI", "CT") ~ 0.5,
                            .default = 0)) %>%
  bind_cols(tibble(y2 = c(7:1) + 0.3)) %>%
  mutate(x2 = case_when(PRO %in% c("MRI") ~ AFF_2,
                       .default = 0)) %>%
  mutate(hjust2 = case_when(PRO %in% c("MRI") ~ 0.5,
                      .default = 0))

label_test_cost
```


```{r}
# affordability

p1_b2 <- df_test_cost %>%
   ggplot()+
    geom_segment(aes(x = AFF_1, xend = AFF_3,
                   y = PRO, yend = PRO),
                 linewidth = 0.4) +
  geom_point(data = df_test_cost_long  %>% filter(AFF_cat %in% c("AFF", "AFF_2")),
             aes(x = AFF_value, y = PRO, colour = AFF_cat),
             size = 1,
             inherit.aes = FALSE)+
  scale_colour_manual(values = c("#1f77b4", "#d62728"),
                    name = "Cost",
                    labels = c("Original",
                               "Out-of-pocket"))+
    geom_text(data = label_test_cost,
            aes(x = x1, y = y1, label = label_1, hjust = hjust1),
             vjust = 0.5, size = 2.8, colour = "black",
            inherit.aes = FALSE)+
  geom_text(data = label_test_cost,
            aes(x = x2, y = y2, label = label_2, hjust = hjust2),
             vjust = 0.5, size = 2.8, colour = "black",
            inherit.aes = FALSE)+
  scale_x_continuous(breaks = seq(0, 2, 0.5),
                     limits = c(0, 2),
                     labels = c("0", "0.5", "1.0", "1.5", "2.0"))+
  coord_cartesian(clip = "off")+
  labs(
       x = "Affordability (%)",
       y = "Procedure")+
  guides(colour = guide_legend(override.aes = list(size = 2.6,
                                                   fill=NA)))+
  theme_classic()+
  theme(plot.title = element_text(size = 12, colour = "black", face = "bold"),
        axis.title.x = element_text(size = 8, colour = "black", face = "bold"),
        axis.title.y = element_blank(),
        axis.text.x = element_text(size = 8, colour = "black"),
        axis.text.y = element_blank(),
        axis.line = element_line(linewidth = 0.3, colour = "black"),
        axis.ticks = element_line(linewidth = 0.3, colour = "black"),
        legend.title = element_blank(),
          #element_text(size = 8, colour = "black", face = "bold"),
        legend.text = element_text(size = 8, colour = "black",
                                   margin = margin(0, 0, 0, 0, unit = "mm")),
        legend.position = "inside",
        legend.position.inside = c(0.55, 0),
        legend.justification.inside = c(0, 0),
        legend.background = element_blank(),
        legend.box.background = element_blank(),
 #       legend.margin = margin(0, 0, 0, 0),
      legend.spacing.x = unit(0, "mm"),
      legend.spacing.y = unit(0, "mm"),
      legend.key = element_rect(fill = NA),
        plot.margin = margin(0, 0 , 0, 0, "mm"))

p1_b2 
```

```{r}
df_test %>%
  relocate(PRO, .after = "PRO_name") %>% View()
```


```{r}
saveRDS(p1_b1, "../plot/p1_b1.RDS")
saveRDS(p1_b2, "../plot/p1_b2.RDS")
```
